+++ tests/WebhookTest.php
@@ -5,7 +5,7 @@ namespace Lloricode\Paymaya\Tests;
 use Carbon\Carbon;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\Psr7\Response;
-use Lloricode\Paymaya\Client\Checkout\WebhookClient;
+use Lloricode\Paymaya\Client\WebhookClient;
 use Lloricode\Paymaya\Request\Webhook\Webhook;
 
 class WebhookTest extends TestCase